+++ pilad/main_test.go
@@ -6,7 +6,6 @@ import (
 	"time"
 )
 
-// TestMain is a hack to get 100% test coverage.
 func TestMain(t *testing.T) {
 	os.Setenv("PILADB_PORT", "35343")
 	go main()
@@ -14,7 +13,15 @@ func TestMain(t *testing.T) {
 	os.Setenv("PILADB_PORT", "")
 }
 
-func TestMainVersion(t *testing.T) {
+func TestMain_Error(t *testing.T) {
+	os.Setenv("PILADB_PORT", "35343")
+	go main()
+	go main()
+	time.Sleep(5 * time.Millisecond)
+	os.Setenv("PILADB_PORT", "")
+}
+
+func TestMain_Version(t *testing.T) {
 	versionFlag = true
 	go main()
 	t.Log(v())